refactor x11 api (w_get_bitmap is dummy)
