+++ src/screenshot.h
@@ -12,9 +12,6 @@
 #define ERRPNGINFO 6
 #define ERRMEMALLOC 7
 
-bitmap_t *x_get_bitmap(int *status);
-char *create_unix_path(char *path, int *status);
-
 int take_screenshot(char *path, bitmap_t *(get_bitmap)(int *));
 
 #endif
